requirements.txt
fastapi
SQLAlchemy
httpx
Pyjwt

uvicorn[standard]
psycopg2-binary
python-dotenv
passlib[bcrypt]
python-multipart
alembic
pydantic[email]
cryptography